chore!: CDK 1.94, naming, eks 1.19
* Updates cdk to 1.94.1
* Unifies naming of resources and fixes scopes of construct resources
* update to eks 1.19

BREAKING CHANGE: Due to naming and scope problem, it's not possible to
upgrade existing super-eks clusters without running into problems.
If you can, it's probably easiest to just create a new stack and
migrate your workloads.
A normal update will try to create a new cluster and delete the old one
which will lead to downtime.

Co-authored-by: